+++ crates/runmat-vm/Cargo.toml
@@ -12,6 +12,7 @@ repository = "https://github.com/runmat-org/runmat"
 runmat-accelerate = { workspace = true, optional = true }
 runmat-accelerate-api = { workspace = true }
 runmat-builtins = { workspace = true }
+runmat-gc = { workspace = true }
 runmat-hir = { workspace = true }
 runmat-parser = { workspace = true }
 runmat-runtime = { workspace = true, default-features = false }
